+++ mobile/library/cc/request_headers_builder.h
@@ -11,6 +11,12 @@
 namespace Envoy {
 namespace Platform {
 
+// Available algorithms to compress requests.
+enum CompressionAlgorithm {
+  Gzip,
+  Brotli,
+};
+
 class RequestHeaders;
 struct RetryPolicy;
 
@@ -22,6 +28,7 @@ public:
 
   RequestHeadersBuilder& addRetryPolicy(const RetryPolicy& retry_policy);
   RequestHeadersBuilder& addUpstreamHttpProtocol(UpstreamHttpProtocol upstream_http_protocol);
+  RequestHeadersBuilder& enableRequestCompression(CompressionAlgorithm algorithm);
 
   RequestHeaders build() const;