+++ mobile/test/swift/EngineBuilderTests.swift
@@ -465,9 +465,7 @@ final class EngineBuilderTests: XCTestCase {
       enableHappyEyeballs: true,
       enableHttp3: http3,
       enableGzipDecompression: true,
-      enableGzipCompression: false,
       enableBrotliDecompression: false,
-      enableBrotliCompression: false,
       enableInterfaceBinding: true,
       enableDrainPostDnsRefresh: false,
       enforceTrustChainVerification: false,
@@ -537,9 +535,7 @@ final class EngineBuilderTests: XCTestCase {
 
     // Compression
     XCTAssertTrue(resolvedYAML.contains(".decompressor.v3.Gzip"))
-    XCTAssertFalse(resolvedYAML.contains(".compressor.v3.Gzip"))
     XCTAssertFalse(resolvedYAML.contains(".decompressor.v3.Brotli"))
-    XCTAssertFalse(resolvedYAML.contains(".compressor.v3.Brotli"))
 
     // Metadata
     XCTAssertTrue(resolvedYAML.contains("device_os: iOS"))
@@ -574,9 +570,7 @@ final class EngineBuilderTests: XCTestCase {
       enableHappyEyeballs: false,
       enableHttp3: false,
       enableGzipDecompression: false,
-      enableGzipCompression: true,
       enableBrotliDecompression: true,
-      enableBrotliCompression: true,
       enableInterfaceBinding: false,
       enableDrainPostDnsRefresh: true,
       enforceTrustChainVerification: true,
@@ -637,9 +631,7 @@ final class EngineBuilderTests: XCTestCase {
 
     // Compression
     XCTAssertFalse(resolvedYAML.contains(".decompressor.v3.Gzip"))
-    XCTAssertTrue(resolvedYAML.contains(".compressor.v3.Gzip"))
     XCTAssertTrue(resolvedYAML.contains(".decompressor.v3.Brotli"))
-    XCTAssertTrue(resolvedYAML.contains(".compressor.v3.Brotli"))
   }
 
   func testReturnsNilWhenUnresolvedValueInTemplate() {
@@ -658,9 +650,7 @@ final class EngineBuilderTests: XCTestCase {
       enableHappyEyeballs: false,
       enableHttp3: false,
       enableGzipDecompression: false,
-      enableGzipCompression: false,
       enableBrotliDecompression: false,
-      enableBrotliCompression: false,
       enableInterfaceBinding: false,
       enableDrainPostDnsRefresh: false,
       enforceTrustChainVerification: true,